aarch64: no breakpoints for okl4 hypervisor
The OKL4 hypervisor does not currently emulate or support
guest use of breakpoint registers. The ID_AA64DFR0_EL.BPRs
and WRPs fields reserve the use of the value zero.

The OKL4 hypervisor uses the value of zero to indicate that there
are no breakpoint registers. This patch adds Linux handling for
zero in these fields.

Signed-off-by: Carl van Schaik <carl@cog.systems>