aarch64: add okl4 paravirt spinlock support
Add support for OKL4 hypervisor paravirtualized spinlocks.

When running as a guest of the OKL4 hypervisor with no affinity,
virtual CPUs may migrate freely to available physical cores, and may
be pre-empted by other vCPUs or VMs.

If a preempted virtual CPU holds a lock, the lock may be held for an
indeterminately long time. Causing the waiter to spin indefinitely.
The OKL4 hypervisor API supports assisting in these cases by
allowing the lock waiter to donate CPU time to the holder, yielding
the waiter to prevent wasted CPU cycles spinning, and rescheduling
the waiter when the lock is released.

Signed-off-by: Carl van Schaik <carl@cog.systems>